+++ exampleCourse/clientFilesCourse/MultiFileClient.js
@@ -1,5 +1,13 @@
 var httpDownloadPrefix = 'data:text/plain;base64,';
 
+// https://developer.mozilla.org/en-US/docs/Web/API/WindowBase64/Base64_encoding_and_decoding#The_Unicode_Problem
+function b64DecodeUnicode(str) {
+    // Going backwards: from bytestream, to percent-encoding, to original string.
+    return decodeURIComponent(atob(str).split('').map(function(c) {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
+    }).join(''));
+}
+
 define(["SimpleClient", "underscore", "clientCode/dropzone"], function(SimpleClient, _, Dropzone) {
     return function(questionTemplate, submissionTemplate) {
         var simpleClient = new SimpleClient.SimpleClient({questionTemplate: questionTemplate, submissionTemplate: submissionTemplate, skipRivets: true});
@@ -63,7 +71,7 @@ define(["SimpleClient", "underscore", "clientCode/dropzone"], function(SimpleCli
                 } else {
                     var $preview = $('<pre><code></code></pre>');
                     try {
-                        var fileContents = atob(fileData);
+                        var fileContents = b64DecodeUnicode(fileData);
                         if (!isBinary(fileContents)) {
                             $preview.find('code').text(fileContents);
                         } else {